requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# AI Providers
google-generativeai==0.3.2
openai==1.3.9

# HTTP and Web
requests==2.31.0
python-multipart==0.0.6

# Voice Features (Optional)
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.13  # Optional - for microphone input

# Configuration and Environment
python-dotenv==1.0.0

# GUI Framework (Built-in with Python)
# tkinter - Built-in with Python, no installation needed
